+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ This document describes changes between each past release.
 3.2.0 (unreleased)
 ==================
 
+- Filter add-ons and plugins in v3 based on the requesting application and version. (#74)
 - Stop exporting cert items to Firefox 58 and above, where they aren't used. (#75)